# Optimize hash single keccak