--- a/risc0/circuit/keccak-sys/kernels/cuda/eval_check.cuh
+++ b/risc0/circuit/keccak-sys/kernels/cuda/eval_check.cuh
@@ -6,10 +6,18 @@
 
 namespace risc0::circuit::keccak::cuda {
 
+extern __device__ FpExt keccak_56(
+    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
 extern __device__ FpExt keccak_55(
     uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
-extern __device__ FpExt keccak_54(
-    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
+extern __device__ FpExt keccak_54(uint32_t idx,
+                                  uint32_t size,
+                                  Fp* arg0,
+                                  FpExt arg1,
+                                  FpExt arg2,
+                                  FpExt arg3,
+                                  FpExt arg4,
+                                  const Fp* arg5);
 extern __device__ FpExt keccak_53(uint32_t idx,
                                   uint32_t size,
                                   Fp* arg0,
@@ -36,12 +44,13 @@ extern __device__ FpExt keccak_51(uint32_t idx,
                                   const Fp* arg5);
 extern __device__ FpExt keccak_50(uint32_t idx,
                                   uint32_t size,
-                                  Fp* arg0,
-                                  FpExt arg1,
+                                  FpExt arg0,
+                                  Fp* arg1,
                                   FpExt arg2,
                                   FpExt arg3,
                                   FpExt arg4,
-                                  const Fp* arg5);
+                                  FpExt arg5,
+                                  const Fp* arg6);
 extern __device__ FpExt keccak_49(uint32_t idx,
                                   uint32_t size,
                                   FpExt arg0,
@@ -53,13 +62,12 @@ extern __device__ FpExt keccak_49(uint32_t idx,
                                   const Fp* arg6);
 extern __device__ FpExt keccak_48(uint32_t idx,
                                   uint32_t size,
-                                  FpExt arg0,
-                                  Fp* arg1,
+                                  Fp* arg0,
+                                  FpExt arg1,
                                   FpExt arg2,
                                   FpExt arg3,
                                   FpExt arg4,
-                                  FpExt arg5,
-                                  const Fp* arg6);
+                                  const Fp* arg5);
 extern __device__ FpExt keccak_47(uint32_t idx,
                                   uint32_t size,
                                   Fp* arg0,
@@ -84,14 +92,8 @@ extern __device__ FpExt keccak_45(uint32_t idx,
                                   FpExt arg3,
                                   FpExt arg4,
                                   const Fp* arg5);
-extern __device__ FpExt keccak_44(uint32_t idx,
-                                  uint32_t size,
-                                  Fp* arg0,
-                                  FpExt arg1,
-                                  FpExt arg2,
-                                  FpExt arg3,
-                                  FpExt arg4,
-                                  const Fp* arg5);
+extern __device__ FpExt keccak_44(
+    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
 extern __device__ FpExt keccak_43(
     uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
 extern __device__ FpExt keccak_42(
@@ -100,20 +102,20 @@ extern __device__ FpExt keccak_41(
     uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
 extern __device__ FpExt keccak_40(
     uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
-extern __device__ FpExt keccak_39(
-    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
+extern __device__ FpExt
+keccak_39(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
 keccak_38(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
 keccak_37(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
 keccak_36(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
-extern __device__ FpExt
-keccak_35(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
+extern __device__ FpExt keccak_35(
+    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
 extern __device__ FpExt keccak_34(
     uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
-extern __device__ FpExt keccak_33(
-    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
+extern __device__ FpExt
+keccak_33(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
 keccak_32(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
@@ -124,8 +126,8 @@ extern __device__ FpExt
 keccak_29(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt keccak_28(
     uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
-extern __device__ FpExt keccak_27(
-    uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, FpExt arg3, const Fp* arg4);
+extern __device__ FpExt
+keccak_27(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
 keccak_26(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
@@ -134,8 +136,15 @@ extern __device__ FpExt
 keccak_24(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
 extern __device__ FpExt
 keccak_23(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
-extern __device__ FpExt
-keccak_22(uint32_t idx, uint32_t size, Fp* arg0, FpExt arg1, FpExt arg2, const Fp* arg3);
+extern __device__ FpExt keccak_22(uint32_t idx,
+                                  uint32_t size,
+                                  Fp* arg0,
+                                  FpExt arg1,
+                                  FpExt arg2,
+                                  FpExt arg3,
+                                  FpExt arg4,
+                                  FpExt arg5,
+                                  const Fp* arg6);
 extern __device__ FpExt keccak_21(uint32_t idx,
                                   uint32_t size,
                                   Fp* arg0,
@@ -152,8 +161,7 @@ extern __device__ FpExt keccak_20(uint32_t idx,
                                   FpExt arg2,
                                   FpExt arg3,
                                   FpExt arg4,
-                                  FpExt arg5,
-                                  const Fp* arg6);
+                                  const Fp* arg5);
 extern __device__ FpExt keccak_19(uint32_t idx,
                                   uint32_t size,
                                   Fp* arg0,
@@ -201,7 +209,8 @@ extern __device__ FpExt keccak_14(uint32_t idx,
                                   FpExt arg2,
                                   FpExt arg3,
                                   FpExt arg4,
-                                  const Fp* arg5);
+                                  FpExt arg5,
+                                  const Fp* arg6);
 extern __device__ FpExt keccak_13(uint32_t idx,
                                   uint32_t size,
                                   Fp* arg0,
@@ -217,7 +226,8 @@ extern __device__ FpExt keccak_12(uint32_t idx,
                                   FpExt arg2,
                                   FpExt arg3,
                                   FpExt arg4,
-                                  const Fp* arg5);
+                                  FpExt arg5,
+                                  const Fp* arg6);
 extern __device__ FpExt keccak_11(uint32_t idx,
                                   uint32_t size,
                                   Fp* arg0,
@@ -233,7 +243,8 @@ extern __device__ FpExt keccak_10(uint32_t idx,
                                   FpExt arg2,
                                   FpExt arg3,
                                   FpExt arg4,
-                                  const Fp* arg5);
+                                  FpExt arg5,
+                                  const Fp* arg6);
 extern __device__ FpExt keccak_9(uint32_t idx,
                                  uint32_t size,
                                  Fp* arg0,
@@ -249,7 +260,8 @@ extern __device__ FpExt keccak_8(uint32_t idx,
                                  FpExt arg2,
                                  FpExt arg3,
                                  FpExt arg4,
-                                 const Fp* arg5);
+                                 FpExt arg5,
+                                 const Fp* arg6);
 extern __device__ FpExt keccak_7(uint32_t idx,
                                  uint32_t size,
                                  Fp* arg0,
@@ -311,7 +323,7 @@ extern __device__ FpExt poly_fp(uint32_t idx,
                                 const Fp* accum);
 
 constexpr size_t INV_RATE = 4;
-constexpr size_t kNumPolyMixPows = 1903;
+constexpr size_t kNumPolyMixPows = 2175;
 extern __constant__ FpExt poly_mix[kNumPolyMixPows];
 
 } // namespace risc0::circuit::keccak::cuda