--- a/risc0/zkvm/src/host/server/exec/tests.rs
+++ b/risc0/zkvm/src/host/server/exec/tests.rs
@@ -1195,6 +1195,16 @@ fn heap_bug_zkvm_527() {
 }
 
 #[test]
+fn keccak_update() {
+    run_test(MultiTestSpec::KeccakUpdate);
+}
+
+#[test]
+fn sha_single_keccak() {
+    run_test(MultiTestSpec::ShaSingleKeccak);
+}
+
+#[test]
 fn sys_keccak_permute() {
     run_test(MultiTestSpec::SysKeccakPermute);
 }