This brings down `sha_single_keccak` from 1843 cycles to 385 cycles. Out of the 385 cycles, 288 are from calling sha256 precompile so we're only adding 89 cycles.